pyproject.toml
[project]
name = "mlsimkit"
description = "Python libraries and commands for training and using ML for CAE use cases" 
dependencies = [
    "setuptools_scm>=8.0",
    "build>=1.2",
    "click>=8.0.0",
    "pydantic>=2.3.0",
    "pyyaml>=6.0",
    "pytest>=8.1",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "sphinx>=7.3",
    "myst-parser>=2.0",
    "mlflow>=2.12.2",
    "numpy~=1.26",
    "pyvista>=0.41",
    "torch>=2.1,<2.6",
    "torch-geometric>=2.4",
    "torch-summary>=1.4",
    "torchmetrics>=1.3",
    "torchvision>=0.16",
    "trimesh>=3.23",
    "tqdm>=4.66.3",
    "vtk>=9.2",
    "scikit-learn>=1.3",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "opencv-python~=4.9.0",
    "accelerate>=0.29",
    "gunicorn>=22.0",           # transitive dependency minimum for pip-audit
    "werkzeug>=3.0.3",          # transitive dependency minimum for pip-audit
    "jinja2>=3.1.4",            # transitive dependency minimum for pip-audit
]
requires-python = ">= 3.9"
readme = "README.md"
dynamic = ["version"]

[project.scripts]
mlsimkit-learn = "mlsimkit.learn.cli:learn"
mlsimkit-accelerate = "mlsimkit.learn.cli:accelerate"
mlsimkit-manifest = "mlsimkit.learn.manifest.cli:manifest_cli"

[build-system]
requires = ["setuptools>=69.0.2", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"mlsimkit" = [
    "conf/*",
    "conf/**/*",            # include all files and subdirs
    "datasets/*",
    "datasets/**/*",        # include all files and subdirs
]

[tool.setuptools_scm]
version_file = "src/mlsimkit/_version.py"
fallback_version = "0.1.0"  # support when .git isn't available 

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tst",
]

[tool.ruff]
target-version = "py39"
line-length = 110
extend-exclude = [
    ".eggs",
]

[tool.bandit]
exclude_dirs = ["tst"]
